refactor(ioc): split up runtime and designtime
