+++ ioc/designtime/types.ts
@@ -1,68 +1,4 @@
-import { IFulfillment, IRequirement, IActivator, IPair } from './interfaces';
-import { DependencyMap } from './container';
-import { metadata } from 'aurelia-metadata';
-import * as AST from './analysis/ast';
-
-export const enum RegistrationFlags {
-  None = 1 << 0,
-  Terminal = 1 << 1
-}
-
-export enum Lifetime {
-  Unspecified,
-  Singleton,
-  Transient
-}
-
-export type DependencyType =
-  | number
-  | Number
-  | string
-  | String
-  | symbol
-  | Symbol
-  | object
-  | Object
-  | Function
-  | AST.INode;
-
-export function validateKey(key: DependencyType): boolean {
-  const type = typeof key;
-
-  return key === 'function' || key === 'object' || key === 'string' || key === 'number';
-}
-
-export function isConstructor(key: DependencyType): boolean {
-  return typeof key === 'function' && Object.prototype.hasOwnProperty.call(key, 'prototype');
-}
-
-export function isASTNode(key: DependencyType): boolean {
-  return key !== null && key !== undefined && (key as AST.INode).isAnalysisASTNode;
-}
-
-export function getParamTypes(ctor: any): any[] {
-  let types: any[] = [];
-  if (ctor.inject === undefined) {
-    types = (metadata.getOwn('design:paramtypes', ctor) as Array<any>) || [];
-  } else {
-    let base = ctor;
-    while (base !== Function.prototype && base !== Object.prototype) {
-      types.push(...getInjectTypes(base));
-      base = Object.getPrototypeOf(base);
-    }
-  }
-  return types;
-}
-
-function getInjectTypes(ctor: any): any[] {
-  if (!Object.prototype.hasOwnProperty.call(ctor, 'inject')) {
-    return [];
-  } else if (typeof ctor.inject === 'function') {
-    return ctor.inject();
-  } else {
-    return ctor.inject;
-  }
-}
+import * as AST from './ast';
 
 export function getTypeSourceFromCtorParameter(param: AST.IParameter): AST.INode | null {
   if (param.parent.kind !== AST.NodeKind.Constructor) {
@@ -162,14 +98,3 @@ function getTypeSourceFromInjectDecorator(param: AST.IParameter): string | null 
 
   return injectDecorator.arguments[param.parent.parameters.indexOf(param)].text;
 }
-
-export class Pair<L, R> implements IPair<L, R> {
-  public readonly left: L;
-  public readonly right: R;
-
-  constructor(left: L, right: R) {
-    this.left = left;
-    this.right = right;
-  }
-}
-
