+++ ioc/runtime/context.ts
@@ -1,8 +1,6 @@
-import { IContext, IRegistration } from "./interfaces";
-import { RegistrationFunctionBuilder, Registration } from "./registration";
-import { DependencyType } from "./types";
-import { IModuleItem } from "./analysis/ast";
-
+import { IContext, IRegistration } from './interfaces';
+import { RegistrationFunctionBuilder, Registration } from './registration';
+import { DependencyType } from './types';
 
 export class DefaultContext implements IContext {
   public readonly builder: RegistrationFunctionBuilder;
@@ -15,7 +13,7 @@ export class DefaultContext implements IContext {
     return new DefaultContext(builder);
   }
 
-  public register<T>(type: DependencyType | IModuleItem): IRegistration<T> {
+  public register<T>(type: DependencyType): IRegistration<T> {
     return new Registration<T>(this, type);
   }
 }
