+++ package.json
@@ -11,9 +11,9 @@
     "karma-watch": "cross-env ./node_modules/.bin/nps karma.watch",
     "karma-debug": "cross-env ./node_modules/.bin/nps karma.watch.debug",
     "build-ioc": "rollup -c ioc.rollup.config.js",
-    "test-ioc-src": "node ioc/test/generate src",
-    "test-ioc-ts-fixture": "node ioc/test/generate test/fixture/ts",
-    "test-ioc-js-fixture": "node ioc/test/generate test/fixture/js",
+    "test-ioc-src": "node ioc/designtime/test/generate src",
+    "test-ioc-ts-fixture": "node ioc/designtime/test/generate test/fixture/ts",
+    "test-ioc-js-fixture": "node ioc/designtime/test/generate test/fixture/js",
     "transpile": "tsc src/compile.ts --outDir scripts",
     "build": "rollup -c compiler.rollup.config.js",
     "test": "node compiler/test/compile",
